# core-banking/api/src/main/resources/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "Core banking",
    "description" : "Socle core banking : comptes, operations, credit, parametrage, arretes, restitutions et etats financiers. Toute reponse porte l'enveloppe data / page / error / meta ; les montants sont des chaines avec leur devise ; les actes a double validation repondent 202 puis s'approuvent par un second porteur."
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/v1/entities/{legalEntityId}/accounts" : {
      "post" : {
        "operationId" : "Account.open",
        "tags" : [ "Account" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Requests.OpenAccount"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "description" : "Soumis a double validation : l'operation en attente, a approuver ou rejeter par un second porteur habilite",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/MakerChecker.View"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/accounts/{accountId}/balance" : {
      "get" : {
        "operationId" : "Account.balance",
        "tags" : [ "Account" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Succes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/AccountUseCases.Balance"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/accounts/{accountId}/blocks" : {
      "post" : {
        "operationId" : "Account.block",
        "tags" : [ "Account" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Requests.BlockAccount"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "description" : "Soumis a double validation : l'operation en attente, a approuver ou rejeter par un second porteur habilite",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/MakerChecker.View"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/accounts/{accountId}/blocks/{blockId}/lift" : {
      "post" : {
        "operationId" : "Account.lift",
        "tags" : [ "Account" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "blockId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Requests.LiftBlock"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "description" : "Soumis a double validation : l'operation en attente, a approuver ou rejeter par un second porteur habilite",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/MakerChecker.View"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/accounts/{accountId}/closure" : {
      "post" : {
        "operationId" : "Account.close",
        "tags" : [ "Account" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Requests.CloseAccount"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "description" : "Soumis a double validation : l'operation en attente, a approuver ou rejeter par un second porteur habilite",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/MakerChecker.View"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
...
    "/v1/entities/{legalEntityId}/accounts/{accountId}/holds" : {
      "post" : {
        "operationId" : "Account.hold",
        "tags" : [ "Account" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Requests.PlaceHold"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "description" : "Soumis a double validation : l'operation en attente, a approuver ou rejeter par un second porteur habilite",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/MakerChecker.View"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/accounts/{accountId}/holds/{holdId}/release" : {
      "post" : {
        "operationId" : "Account.release",
        "tags" : [ "Account" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "holdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "202" : {
            "description" : "Soumis a double validation : l'operation en attente, a approuver ou rejeter par un second porteur habilite",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/MakerChecker.View"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/accounts/{accountId}/journal" : {
      "get" : {
        "operationId" : "Account.journal",
        "tags" : [ "Account" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "$ref" : "#/components/parameters/Page"
        }, {
          "$ref" : "#/components/parameters/PageSize"
        } ],
        "responses" : {
          "200" : {
            "description" : "Succes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "type" : "array",
                      "items" : {
                        "$ref" : "#/components/schemas/Journal.StatementLine"
                      }
                    },
                    "page" : {
                      "$ref" : "#/components/schemas/Page"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/accounts/{accountId}/ledger" : {
      "get" : {
        "operationId" : "Account.ledger",
        "tags" : [ "Account" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "$ref" : "#/components/parameters/After"
        }, {
          "$ref" : "#/components/parameters/PageSize"
        } ],
        "responses" : {
          "200" : {
            "description" : "Succes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "type" : "array",
                      "items" : {
                        "$ref" : "#/components/schemas/Journal.StatementLine"
                      }
                    },
                    "page" : {
                      "$ref" : "#/components/schemas/Page"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "description" : "Jeton d'acces Keycloak : l'entite, l'agence et les roles de l'appelant en sont derives, jamais d'un parametre.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT",
        "type" : "http"
      }
    },
    "parameters" : {
...
      "RequestId" : {
        "name" : "X-Request-Id",
        "in" : "header",
        "required" : false,
        "schema" : {
          "type" : "string",
          "pattern" : "^[A-Za-z0-9._:-]{1,64}$"
        },
        "description" : "Identifiant de requete du client, repris dans meta.requestId et en en-tete de reponse ; attribue s'il est absent ou mal forme"
      }
    },
    "responses" : {
      "400" : {
        "description" : "Requete invalide : page hors bornes, parametre mal forme, cle d'idempotence absente, corps illisible, curseur invalide",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "401" : {
        "description" : "Aucun jeton valide",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas" : {
...
      "MakerChecker.View" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "operation" : {
            "type" : "string"
          },
          "handler" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "makerId" : {
            "type" : "string"
          },
          "makerUsername" : {
            "type" : "string"
          },
          "madeAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "decidedBy" : {
            "type" : "string"
          },
          "decidedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "decisionReason" : {
            "type" : "string"
          },
          "payload" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "result" : { },
          "error" : {
            "type" : "string"
          }
        }
      },
      "Meta" : {
        "type" : "object",
        "properties" : {
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "requestId" : {
            "type" : "string"
          }
        },
        "required" : [ "timestamp", "requestId" ]
      },
...
      "Requests.OpenAccount" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "holderPartyId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "productCode" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}